댓글 기능 구현

+++ app/Http/Controllers/ForumController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Category;
 use App\Models\Post;
+use App\Models\Reply;
 
 class ForumController extends Controller
 {
@@ -30,16 +31,22 @@ class ForumController extends Controller
     }
     public function create()
     {
-        $categories = Category::orderby('title', 'asc')->get();
-        return view('forum.create')->with('categories', $categories);
+        if(isset(auth()->user()->id)){
+            $categories = Category::orderby('title', 'asc')->get();
+            return view('forum.create')->with('categories', $categories);
+        }else{
+            return redirect('/');
+        }
     }
     public function store(Request $request)
     {
-        $post = new Post;
-        $post->title = $request->title;
-        $post->category_id = $request->category_id;
-        $post->content = $request->content;
-        $post->save();
+        if(isset(auth()->user()->id)){
+            $post = new Post;
+            $post->title = $request->title;
+            $post->category_id = $request->category_id;
+            $post->content = $request->content;
+            $post->save();
+        }
 
         $result = $request->all();
 
@@ -51,11 +58,13 @@ class ForumController extends Controller
     }
     public function update(Request $request)
     {
-        $post = Post::find($request->post_id);
-        $post->title = $request->title;
-        $post->category_id = $request->category_id;
-        $post->content = $request->content;
-        $post->save();
+        if(isset(auth()->user()->id)){
+            $post = Post::find($request->post_id);
+            $post->title = $request->title;
+            $post->category_id = $request->category_id;
+            $post->content = $request->content;
+            $post->save();
+        }
 
         $result = $request->all();
 
@@ -76,8 +85,21 @@ class ForumController extends Controller
     public function delete($id)
     {
         $post = Post::find($id);
-        $post->delete();
+        if(auth()->user()->id == $post->user_id){
+            $post->delete();
+        }
 
         return redirect('/');
+    }
+    public function replyStore(Request $request)
+    {
+        $reply = new Reply;
+        $reply->user_id = auth()->user()->id;
+        $reply->post_id = $request->post_id;
+        $reply->reply = $request->reply;
+        $reply->save();
+
+        return redirect('/'.$request->post_id.'/view');
+
     }
 }
